# apps/starter-kit/api/swagger.yaml
openapi: 3.0.3
info:
  title: Hackathon - Starter Kit
  description: >
    This Hackathon starter kit is designed to give teams a quick head start 
    on building integrations with Open Finance APIs.
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /token/client-credentials:
    post:
      summary: Obtain client credentials token
      tags:
        - Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scope
              properties:
                scope:
                  type: string
                  enum:
                    - confirmation-of-payee
                    - openid payments
                    - openid accounts
                    - openid products
              example:
                scope: "confirmation-of-payee"
      responses:
        '200':
          description: Client credentials token issued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
                  scope:
                    type: string
        '400':
          description: Invalid request due to missing or unsupported scope
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                    example: "Invalid scope"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /token/authorization-code:
    post:
      summary: Exchange authorization code for tokens
      tags:
        - Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - code_verifier
              properties:
                code:
                  type: string
                code_verifier:
                  type: string
              example:
                code: "abc123authorizationcode"
                code_verifier: "xyz456codeverifier"
      responses:
        '200':
          description: Tokens issued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
                  scope:
                    type: string
        '400':
          description: Missing or invalid code/code_verifier
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                    example: "code and code_verifier required"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /token/refresh-token:
    post:
      summary: Refresh access token using a refresh token
      tags:
        - Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
              example:
                refresh_token: "xyz456refreshtoken"
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
                  scope:
                    type: string
        '400':
          description: Missing or invalid refresh_token
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                    example: "refresh_token required"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...